requirements.txt
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
dvc>=3.0.0
mlflow>=2.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.32.0
pyyaml>=6.0
